added sekecto directory
