+++ 3142A_ELEVATED.v5code
@@ -1,1 +1,1 @@
-{"title":"3142A_ELEVATED","description":"Empty V5 C++ Project","icon":"USER921x.bmp","version":"20.02.1421","sdk":"20200618_17_00_00","language":"cpp","competition":false,"files":[{"name":"include/robot-config.h","type":"File","specialType":"device_config"},{"name":"include/vex.h","type":"File","specialType":""},{"name":"include/pid.h","type":"File","specialType":""},{"name":"include/mathAndConstants.h","type":"File","specialType":""},{"name":"include/chassisfunctions.h","type":"File","specialType":""},{"name":"include/selectorsource.h","type":"File","specialType":""},{"name":"include/motionprofile.h","type":"File","specialType":""},{"name":"include/macros.h","type":"File","specialType":""},{"name":"include/literals.h","type":"File","specialType":""},{"name":"include/usercontrol.h","type":"File","specialType":""},{"name":"include/globals.h","type":"File","specialType":""},{"name":"include/premacros.h","type":"File","specialType":""},{"name":"makefile","type":"File","specialType":""},{"name":"src/main.cpp","type":"File","specialType":""},{"name":"src/robot-config.cpp","type":"File","specialType":"device_config"},{"name":"src/pid.cpp","type":"File","specialType":""},{"name":"src/mathAndConstants.cpp","type":"File","specialType":""},{"name":"src/chassisfunctions.cpp","type":"File","specialType":""},{"name":"src/selectorsource.cpp","type":"File","specialType":""},{"name":"src/motionprofile.cpp","type":"File","specialType":""},{"name":"src/macros.cpp","type":"File","specialType":""},{"name":"src/literals.cpp","type":"File","specialType":""},{"name":"src/globals.cpp","type":"File","specialType":""},{"name":"src/usercontrol.cpp","type":"File","specialType":""},{"name":"vex/mkenv.mk","type":"File","specialType":""},{"name":"vex/mkrules.mk","type":"File","specialType":""},{"name":"README.md","type":"File","specialType":""},{"name":"include","type":"Directory"},{"name":"src","type":"Directory"},{"name":"vex","type":"Directory"}],"device":{"slot":3,"uid":"276-4810","options":{}},"isExpertMode":true,"isExpertModeRC":true,"isVexFileImport":false,"robotconfig":[],"neverUpdate":null}
+{"title":"3142A_ELEVATED","description":"Empty V5 C++ Project","icon":"USER921x.bmp","version":"20.02.1421","sdk":"20200618_17_00_00","language":"cpp","competition":false,"files":[{"name":"include/robot-config.h","type":"File","specialType":"device_config"},{"name":"include/vex.h","type":"File","specialType":""},{"name":"include/pid.h","type":"File","specialType":""},{"name":"include/mathAndConstants.h","type":"File","specialType":""},{"name":"include/chassisfunctions.h","type":"File","specialType":""},{"name":"include/selectorsource.h","type":"File","specialType":""},{"name":"include/motionprofile.h","type":"File","specialType":""},{"name":"include/macros.h","type":"File","specialType":""},{"name":"include/literals.h","type":"File","specialType":""},{"name":"include/usercontrol.h","type":"File","specialType":""},{"name":"include/globals.h","type":"File","specialType":""},{"name":"include/premacros.h","type":"File","specialType":""},{"name":"include/Selector/selectorAPI.h","type":"File","specialType":""},{"name":"include/Selector/selectorImpl.h","type":"File","specialType":""},{"name":"makefile","type":"File","specialType":""},{"name":"src/main.cpp","type":"File","specialType":""},{"name":"src/robot-config.cpp","type":"File","specialType":"device_config"},{"name":"src/pid.cpp","type":"File","specialType":""},{"name":"src/mathAndConstants.cpp","type":"File","specialType":""},{"name":"src/chassisfunctions.cpp","type":"File","specialType":""},{"name":"src/selectorsource.cpp","type":"File","specialType":""},{"name":"src/motionprofile.cpp","type":"File","specialType":""},{"name":"src/macros.cpp","type":"File","specialType":""},{"name":"src/literals.cpp","type":"File","specialType":""},{"name":"src/globals.cpp","type":"File","specialType":""},{"name":"src/usercontrol.cpp","type":"File","specialType":""},{"name":"src/Selector/selectorAPI.cpp","type":"File","specialType":""},{"name":"src/Selector/selectorImpl.cpp","type":"File","specialType":""},{"name":"vex/mkenv.mk","type":"File","specialType":""},{"name":"vex/mkrules.mk","type":"File","specialType":""},{"name":"README.md","type":"File","specialType":""},{"name":"include","type":"Directory"},{"name":"include/Selector","type":"Directory"},{"name":"src","type":"Directory"},{"name":"src/Selector","type":"Directory"},{"name":"vex","type":"Directory"}],"device":{"slot":3,"uid":"276-4810","options":{}},"isExpertMode":true,"isExpertModeRC":true,"isVexFileImport":false,"robotconfig":[],"neverUpdate":null}
